feat: structure reorganized
