--- product-service/src/models/product.model.ts
@@ -1,13 +1,0 @@
-export enum SKATE_TYPES {
-  DERBY = 'Roller derby skates',
-  RETRO = 'Retro Skates & Disco Roller'
-}
-
-export interface IProduct {
-  id?: string;
-  title: string;
-  description?: SKATE_TYPES;
-  price: number;
-  image?: string;
-  count?: number;
-}
