+++ src/product-service/db/constants.ts
