+++ src/product-service/handler.ts
@@ -1,0 +1,11 @@
+import { getProductsById } from './src/handlers/get-product-by-id';
+import { getProductsList } from './src/handlers/get-products-list';
+import { postProduct } from './src/handlers/post-product';
+import { invoke } from './db/pg-client-lambda';
+
+export {
+  getProductsList,
+  getProductsById,
+  postProduct,
+  invoke
+}
