+++ src/product-service/jest.config.js
