+++ src/product-service/src/constants/constants.ts
