+++ src/product-service/src/models/product.model.ts
@@ -1,0 +1,8 @@
+export interface IProduct {
+  id?: string;
+  title: string;
+  description?: string;
+  price: number;
+  image?: string;
+  count?: number;
+}
