+++ src/product-service/src/models/stoke.model.ts
