+++ src/product-service/webpack.config.js
